Fixed problem with db utc date. Removed unused note controller.
