--- OsmIntegrator/ApiModels/AddNote.cs
@@ -1,19 +1,0 @@
-using System;
-using System.ComponentModel.DataAnnotations;
-
-public class NewNote
-{
-    public Guid UserId { get; set; }
-
-    [Required]
-    public double? Lat { get; set; }
-
-    [Required]
-    public double? Lon { get; set; }
-
-    [Required]
-    public Guid? TileId { get; set; }
-
-    public string Text { get; set; }
-
-}
