--- OsmIntegrator/ApiModels/MessageInput.cs
@@ -1,21 +1,0 @@
-using System;
-using System.ComponentModel.DataAnnotations;
-
-namespace OsmIntegrator.ApiModels
-{
-  public class MessageInput
-  {
-    public Guid? ConversationId { get; set; }
-
-    public string Text { get; set; }
-
-    public double? Lat { get; set; }
-
-    public double? Lon { get; set; }
-
-    public Guid? StopId { get; set; }
-
-    [Required]
-    public Guid? TileId { get; set; }
-  }
-}
